requirements.txt
Django==4.1.4
django-admin-rangefilter==0.9.0
django-extensions==3.2.1
django-money==3.0.0
python-slugify==8.0.0